+++ packages/plugin-rax-web/src/index.ts
@@ -14,7 +14,7 @@ import getAppEntry from './utils/getAppEntry';
 
 const { getMpaEntries } = appHelpers;
 export default (api) => {
-  const { onGetWebpackConfig, getValue, context, registerTask, registerCliOption, applyMethod, onHook } = api;
+  const { onGetWebpackConfig, getValue, context, registerTask, registerCliOption, applyMethod } = api;
 
   const getWebpackBase = getValue(GET_RAX_APP_WEBPACK_CONFIG);
   const tempDir = getValue('TEMP_PATH');
@@ -45,9 +45,6 @@ export default (api) => {
     }
   });
 
-  // Web entries
-  let entries = [getAppEntry(rootDir)];
-
   onGetWebpackConfig(target, (config) => {
     // eslint-disable-next-line @typescript-eslint/no-shadow
     const { command } = context;
@@ -58,7 +55,7 @@ export default (api) => {
         api,
         staticConfig,
         documentPath,
-        pages: webConfig.mpa ? [] : entries,
+        pages: webConfig.mpa ? [] : [getAppEntry(rootDir)],
         target,
       },
     ]);
@@ -71,10 +68,6 @@ export default (api) => {
     }
 
     if (webConfig.mpa || webConfig.pha) {
-      entries = getMpaEntries(api, {
-        target,
-        appJsonContent: staticConfig,
-      });
       // support --mpa-entry to specify mpa entry
       registerCliOption({
         name: 'mpa-entry',
@@ -84,7 +77,10 @@ export default (api) => {
         type: 'web',
         framework: 'rax',
         targetDir: tempDir,
-        entries,
+        entries: getMpaEntries(api, {
+          target,
+          appJsonContent: staticConfig,
+        }),
       });
     }
 
@@ -102,24 +98,6 @@ export default (api) => {
     }
     setLocalBuilder(api);
   }
-
-  // Remove comment node
-  let webBuildDir;
-
-  onHook('before.build.run', ({ config: configs }) => {
-    const config = configs.find((configItem) => configItem.name === 'web');
-    webBuildDir = config.output.path;
-  });
-
-  onHook('after.build.compile', () => {
-    entries.forEach(({ entryName }) => {
-      const htmlFilePath = path.join(webBuildDir, `${entryName}.html`);
-      if (fs.existsSync(htmlFilePath)) {
-        const html = fs.readFileSync(htmlFilePath, 'utf-8');
-        fs.writeFileSync(htmlFilePath, html.replace(/(\<\!--__INNER_ROOT__--\>|\<\!--__BEFORE_ROOT__--\>|\<\!--__AFTER_ROOT__--\>)/g, ''));
-      }
-    });
-  });
 };
 
 function getAbsolutePath(filepath: string): string | undefined {