+++ packages/plugin-rax-web/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v2.0.4
 
 - Fix: not filter routes by target in document plugin
+- Fix: DEF remove comment node check
 
 ## v2.0.3