+++ packages/plugin-rax-web/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v2.0.4
+
+- Fix: not filter routes by target in document plugin
+- Fix: DEF remove comment node check
+
 ## v2.0.3
 
 - Chore: add logic protect